fix(ci): --all-features clippy, deprecated API, unused imports
- Fixed SyncConfig import path in watch.rs and watch_session.rs
- Fixed EndpointPair::Local construction in watch_session.rs (Box<dyn Endpoint>)
- Added mode field to FileEntry in s3.rs and gcs.rs placeholder endpoints
- Fixed deprecated TempDir::into_path → keep() in filter tests
- Fixed unused MetadataExt imports in edge_cases and basic tests
- Fixed unnecessary u64 cast in protocol.rs
- Fixed set_permissions usage in basic.rs test
- Removed unused SymlinkMode/ChecksumType imports from watch.rs

+++ src/streaming/protocol.rs
@@ -1621,7 +1621,7 @@ mod tests {
             checksums.push(BlockChecksum {
                 offset: i * 4096,
                 weak: i as u32,
-                strong: i as u64,
+                strong: i,
             });
         }
         let entry = DestFileEntry {
